separate Helm repo functions, persist Helm repos to db
