+++ helm/model.go
@@ -1,0 +1,74 @@
+// Copyright © 2019 Banzai Cloud
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+package helm
+
+import (
+	"fmt"
+	"strings"
+	"time"
+
+	"github.com/jinzhu/gorm"
+	"github.com/sirupsen/logrus"
+)
+
+// TableName constants
+const (
+	helmRepoTableName = "helmrepositories"
+)
+
+// represents a collection of parameters for chart repository
+type HelmRepoModel struct {
+	ID        uint   `gorm:"primary_key"`
+	Name      string `gorm:"unique_index:idx_unique_id"`
+	OrgID     uint   `gorm:"unique_index:idx_unique_id"`
+	CreatedAt time.Time
+	UpdatedAt time.Time
+	DeletedAt *time.Time `gorm:"unique_index:idx_unique_id" sql:"index"`
+	CreatedBy uint
+	URL       string
+	Username  string
+	Password  string
+	CertFile  string `sql:"type:text"`
+	KeyFile   string `sql:"type:text"`
+	CAFile    string `sql:"type:text"`
+}
+
+// TableName changes the default table name.
+func (HelmRepoModel) TableName() string {
+	return helmRepoTableName
+}
+
+// String method prints formatted cluster fields.
+func (g HelmRepoModel) String() string {
+	return fmt.Sprintf("Name: %d, Creation date: %s, URL: %s", g.Name, g.CreatedAt, g.URL)
+}
+
+// Migrate executes the table migrations for the cluster module.
+func Migrate(db *gorm.DB, logger logrus.FieldLogger) error {
+	tables := []interface{}{
+		&HelmRepoModel{},
+	}
+
+	var tableNames string
+	for _, table := range tables {
+		tableNames += fmt.Sprintf(" %s", db.NewScope(table).TableName())
+	}
+
+	logger.WithFields(logrus.Fields{
+		"table_names": strings.TrimSpace(tableNames),
+	}).Info("migrating model tables")
+
+	return db.AutoMigrate(tables...).Error
+}
